Backend 3 type of serializers separated and all descriptions and note provided in api documentation
