+++ backend/Users/urls.py
@@ -1,7 +1,7 @@
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
 from config.router import Router
-from .views import  InstructorViewSet, StudentViewSet,EnrollmentViewSet,  FileCategoryViewSet, UserTypeViewSet, FilesViewSet, BillViewSet,InstructorHomeAPI,StudentHomeAPI
+from .views import  InstructorViewSet, StudentViewSet,EnrollmentViewSet,  FileCategoryViewSet, UserTypeViewSet, FilesViewSet, BillViewSet
 router = DefaultRouter()
 
 router.register(f'{Router["student_account"]["children"]["instructor"]}', InstructorViewSet, basename='instructor')
@@ -15,7 +15,5 @@ router.register(f'{Router["student_account"]["children"]["bill"]}', BillViewSet,
 
 urlpatterns = [
     path("",include(router.urls)),
-    path("ihome/<str:id>/",InstructorHomeAPI.as_view()),
-    path("shome/<str:id>/",StudentHomeAPI.as_view()),
 
 ]
