+++ backend/Users/views.py
@@ -1,14 +1,11 @@
-from collections import defaultdict
 from rest_framework import viewsets
-from django.db.models import Sum, Count
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from scheduling.models import Appointment,TimeSlot
 from django.shortcuts import render
 from .models import  Instructor,Student,Enrollment,FileCategory,UserType,Files,Bill
 from .serializer import InstructorSerializer, StudentSerializer,  \
-       EnrollmentSerializer, FileCategorySerializer,  UserTypeSerializer, FilesSerializer, BillSerializer,\
-    AppointmentSerializer_,EnrollmentSerializer_,TimeSlotSerializer_
+       EnrollmentSerializer, FileCategorySerializer,  UserTypeSerializer, FilesSerializer, BillSerializer
 
 # Create your views here.
 
@@ -26,7 +23,6 @@ class StudentViewSet(viewsets.ModelViewSet):
         serializer.save()
     def perform_update(self, serializer):
         serializer.save()
-
 class EnrollmentViewSet(viewsets.ModelViewSet):
     queryset = Enrollment.objects.all()
     serializer_class = EnrollmentSerializer
@@ -34,7 +30,6 @@ class EnrollmentViewSet(viewsets.ModelViewSet):
         serializer.save()
     def perform_update(self, serializer):
         serializer.save()
-
 class FileCategoryViewSet(viewsets.ModelViewSet):
     queryset = FileCategory.objects.all()
     serializer_class = FileCategorySerializer
@@ -42,7 +37,6 @@ class FileCategoryViewSet(viewsets.ModelViewSet):
         serializer.save()
     def perform_update(self, serializer):
         serializer.save()
-
 class UserTypeViewSet(viewsets.ModelViewSet):
     queryset = UserType.objects.all()
     serializer_class = UserTypeSerializer
@@ -64,62 +58,5 @@ class BillViewSet(viewsets.ModelViewSet):
         serializer.save()
     def perform_update(self, serializer):
         serializer.save()
-class EnrollmentViewSet(viewsets.ModelViewSet):
-    queryset = Enrollment.objects.all()
-    serializer_class = EnrollmentSerializer
-    def perform_create(self, serializer):
-        serializer.save()
-    def perform_update(self, serializer):
-        serializer.save()
 
 
-
-class BillStatisticsByType(APIView):
-    def get(self,request):
-        data = Bill.objects.values("package", "type").annotate(
-            total_amount=Sum("price"), count=Count("package")
-        )
-        ready_data = defaultdict(lambda: {"Count": 0})
-        for bill in data:
-            package = bill["package"]
-            type = bill["type"]
-            price = bill["total_amount"]
-            count = bill["count"]
-            ready_data[package][type] = price
-            ready_data[package]["Count"] += count
-        ready_data = dict(ready_data)
-
-        return Response(ready_data)
-
-class InstructorHomeAPI(APIView):
-    def get(self,request,id):
-        ready_data = {}
-        appointments  = Appointment.objects.filter(time_slot__staff_id=id)
-        serializer = AppointmentSerializer_(appointments,many=True)
-        for i in serializer.data:
-            i["time_slot"] = TimeSlotSerializer_(TimeSlot.objects.get(id=i["time_slot"])).data
-
-        return Response(serializer.data)
-
-class StudentHomeAPI(APIView):
-    def get(self,request,id):
-        enrolment = Enrollment.objects.filter(student__id=id)
-        print(enrolment)
-        enrolment = EnrollmentSerializer_(enrolment,many=True)
-        bill = Bill.objects.filter(student__id=id)
-        bill = BillSerializer(bill,many=True)
-        files = Files.objects.filter(student__id=id)
-        files = FilesSerializer(files,many=True)
-        appointments  = Appointment.objects.filter(student__id=id)
-        appointments = AppointmentSerializer_(appointments,many=True)
-        for i in enrolment.data:
-            i["bill"] = bill.data
-            i["files"]= files.data
-            i["appointments"] = appointments.data
-
-        return Response(enrolment.data)
-
-
-
-def landing_page(request):
-    return render(request, 'index.html')
