+++ backend/config/router.py
@@ -14,9 +14,8 @@ Router = {
     "account_management":{
         "base":"account_management/",
         "methods":["POST","GET","UPDATE","DELETE"],
-        "descriptions":"Here we provide simple URLs to represent the communication part of our website",
+        "descriptions":"Here we provide simple URLs to represent the account management part of our website",
         "note":"If you access this API, you will receive all data. Adding <span>ID</span> after the endpoint will retrieve specific data of that model.",
-
         "children":{
             "schools":"schools",
             "class":"class",
@@ -29,9 +28,8 @@ Router = {
     "services":{
         "base":"account_management/services/",
         "methods":["POST","GET","UPDATE","DELETE"],
-        "descriptions":"Here we provide simple URLs to represent the communication part of our website",
-        "note":"If you access this API, you will receive all data. Adding <span>ID</span> after the endpoint will retrieve specific data of that model.",
-
+        "descriptions":"Here we provide simple URLs to represent the services part of our website. And it is child of account management. because of its size we divided it to 2 parts",
+        "note":"Attempting to reach <span>services/</span> endpoint will be result with <span class='highlight-red'>ERROR</span>",
         "children":{
             "add_on":"add_on",
             "component":"component",
@@ -47,9 +45,8 @@ Router = {
     "configuration": {
         "base": "configuration/",
         "methods":["POST","GET","UPDATE","DELETE"],
-        "descriptions":"Here we provide simple URLs to represent the communication part of our website",
+        "descriptions":"Here we provide simple URLs to represent the configuration part of our website",
         "note":"If you access this API, you will receive all data. Adding <span>ID</span> after the endpoint will retrieve specific data of that model.",
-
         "children": {
             "company":"company",
             "zipcode":"zipcode",
@@ -69,7 +66,7 @@ Router = {
     "student_account":{
         "base":"student_account/",
         "methods":["POST","GET","UPDATE","DELETE"],
-        "descriptions":"Here we provide simple URLs to represent the communication part of our website",
+        "descriptions":"Here we provide simple URLs to represent the student account part of our website",
         "note":"If you access this API, you will receive all data. Adding <span>ID</span> after the endpoint will retrieve specific data of that model.",
 
         "children":{
@@ -84,9 +81,9 @@ Router = {
     },
     "main_admin":{
         "base":"main_admin/",
-        "methods":["POST","GET","UPDATE","DELETE"],
-        "descriptions":"Here we provide simple URLs to represent the communication part of our website",
-        "note":"If you access this API, you will receive all data. Adding <span>ID</span> after the endpoint will retrieve specific data of that model.",
+        "methods":["GET","REST METHODS ARE BLOCKED"],
+        "descriptions":"This api represents domain and database schema name of each website. This side only can be changed via backend and super-admin panel",
+        "note":"Any attempt to <span>POST</span>  <span>DELETE</span>  <span>UPDATE</span> will result with  <span class='highlight-red'>ERROR</span>",
 
         "children":{
 
@@ -94,21 +91,30 @@ Router = {
     },
     "statistics":{
         "base":"statistics/",
-        "methods":["POST","GET","UPDATE","DELETE"],
-        "descriptions":"Here we provide simple URLs to represent the communication part of our website",
-        "note":"If you access this API, you will receive all data. Adding <span>ID</span> after the endpoint will retrieve specific data of that model.",
+        "methods":["GET","REST METHODS ARE BLOCKED"],
+        "descriptions":"This api automatically generated by backend for this reason there only get method. ",
+        "note":"Any attempt to <span>POST</span>  <span>DELETE</span>  <span>UPDATE</span> will result with  <span class='highlight-red'>ERROR</span>",
 
         "children":{
             "expanses":"expanses",
             "bill":"bill",
         }
     },
+    "page_api":{
+        "base":"page_api/",
+        "methods":["GET","REST METHODS ARE BLOCKED"],
+        "descriptions":"Here we provide simple URLs to get page base api this is longest api and here you can create or delete. This api designed only for get method",
+        "note":"Any attempt to <span>POST</span>  <span>DELETE</span>  <span>UPDATE</span> will result with  <span class='highlight-red'>ERROR</span>",
+        "children":{
+            "instructor":"instructor/",
+            "student":"student/"
+        }
+    },
     "scheduling":{
         "base":"scheduling/",
         "methods":["POST","GET","UPDATE","DELETE"],
-        "descriptions":"Here we provide simple URLs to represent the communication part of our website",
+        "descriptions":"Here we provide simple URLs to represent the scheduling part of our website",
         "note":"If you access this API, you will receive all data. Adding <span>ID</span> after the endpoint will retrieve specific data of that model.",
-
         "children":{
             "time_slot":"time_slot",
             "week_range":"week_range",
@@ -120,10 +126,9 @@ Router = {
     },
     "abstracts":{
         "base":"",
-        "methods":["POST","GET","UPDATE","DELETE"],
-        "descriptions":"Here we provide simple URLs to represent the communication part of our website",
-        "note":"If you access this API, you will receive all data. Adding <span>ID</span> after the endpoint will retrieve specific data of that model.",
-
+        "methods":["ALL METHODS ARE BLOCKED"],
+        "descriptions":"Here we provide abstract classes because of abstractness of this model we can not provide api for as for this reason mainly this block will be empty",
+        "note":"If you request to get <span>abstracts/</span> you will get <span class='highlight-red'>ERROR</span>",
         "children":{
             "":""
         }
