+++ backend/templates/Doccumentation.html
@@ -82,18 +82,18 @@
             color: #999;
             font-size: 14px;
         }
-        .highlight-red {
+        .sp-wr .highlight-red {
             color: #ff5252;
             font-weight: bold;
         }
-        .highlight-red, span{
+        .sp-wr span:not(.highlight-red){
 
             background-color: #555;
             padding: 8px 12px;
             border-radius: 5px;
             font-family: 'Courier New', Courier, monospace;
             font-size: 14px;
-            color: #ddd;
+            color: #ffb7b7;
             margin-bottom: 20px;
             display: inline-block;
 
@@ -139,7 +139,7 @@
                 </div>
                 <div class="note">
                     {% if keys.note != None %}
-                        <span class="highlight-red">REMEMBER:</span> {{keys.note|safe}}
+                        .<div class="sp-wr"> <span class="highlight-red">REMEMBER:</span> {{keys.note|safe}}</div>
                     {% endif %}
                 </div>
             </div>
